batched affine formulas for TE - too expensive
